memory

+++ src/io/puts.c
@@ -3,11 +3,12 @@
 #include <errno.h>
 
 int puts(const char *s) {
-    long ret = syscall(1, 1, s, strlen(s));
+    int ret = syscall(1, 1, s, strlen(s));
 
     if (ret < 0) {
         errno = -ret;
         return -1;
     }
+    
     return ret;
 }
